+++ HISTORY.md
@@ -1,3 +1,15 @@
+v1.6.4
+=================
+* added mega demo, fixed spelling mistake in ghp.sh
+
+v1.6.3
+=================
+* Updated vulcanize index and demo
+
+v1.6.2
+=================
+* Added vulcanize
+
 v1.6.1
 =================
 * Load moment from px-datetime-imports